% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-weight.R
\name{get_weight}
\alias{get_weight}
\title{Get weight}
\usage{
get_weight(model, n_gram, pos = 1L, time = 0,
  update_excluded = FALSE, zero_indexed = FALSE)
}
\arguments{
\item{model}{A PPM model object as produced by (for example)
\code{\link{new_ppm_simple}} or \code{\link{new_ppm_decay}}.}

\item{n_gram}{An integerish vector defining the n-gram to be queried.}

\item{pos}{(Integerish scalar)
The nominal 'position' at which the n-gram is retrieved
(only relevant for decay-based models).}

\item{time}{(Numeric scalar)
The nominal 'time' at which the n-gram is retrieved
(only relevant for decay-based models).}

\item{update_excluded}{(Logical scalar)
Whether to retrieve update-excluded counts or not.}

\item{zero_indexed}{(Logical scalar)
Whether the input n-gram is zero-indexed.}
}
\value{
A numeric scalar identifying the weight of the n-gram.
}
\description{
Gets the weight (or count) of a given n-gram in a trained PPM moel.
}
