% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{is_ppm}
\alias{is_ppm}
\title{Is 'x' a 'ppm' object?}
\usage{
is_ppm(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
TRUE if the object is of class "ppm", FALSE otherwise.
}
\description{
Tests for objects of class "ppm".
}
