% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{is_ppm_decay}
\alias{is_ppm_decay}
\title{Is 'x' a 'ppm_decay' object?}
\usage{
is_ppm_decay(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
TRUE if the object is of class "ppm_decay", FALSE otherwise.
}
\description{
Tests for objects of class "ppm_decay".
}
