% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{is_ppm_simple}
\alias{is_ppm_simple}
\title{Is 'x' a 'ppm_simple' object?}
\usage{
is_ppm_simple(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
TRUE if the object is of class "ppm_simple", FALSE otherwise.
}
\description{
Tests for objects of class "ppm_simple".
}
