% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-seq.R
\name{model_seq}
\alias{model_seq}
\title{Model sequence}
\usage{
model_seq(model, seq, time = NULL, train = TRUE, predict = TRUE,
  return_distribution = TRUE, return_entropy = TRUE)
}
\value{
a \code{\link[tibble]{tibble}}.
}
\description{
Models a sequence using a PPM model.
}
