% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{new_ppm_decay}
\alias{new_ppm_decay}
\title{Create decay-based PPM model}
\usage{
new_ppm_decay(alphabet_size, order_bound = 10L, ltm_weight = 1,
  ltm_half_life = 10, ltm_asymptote = 0, noise = 0, stm_weight = 1,
  stm_duration = 0, buffer_weight = 1, buffer_length_time = 0,
  buffer_length_items = 0L, only_learn_from_buffer = FALSE,
  only_predict_from_buffer = FALSE,
  seed = sample.int(.Machine$integer.max, 1))
}
\arguments{
\item{alphabet_size}{(Integerish scalar)
The size of the alphabet from which sequences are drawn.}

\item{order_bound}{(Integerish scalar)
The model's Markov order bound.}

\item{noise}{(Numeric scalar)
Scale parameter for the retrieval noise distribution.}

\item{stm_weight}{(Numeric scalar)
The weight of n-grams immediately after leaving the buffer.}

\item{buffer_weight}{(Numeric scalar)
The weight of n-grams within the buffer.}

\item{buffer_length_time}{(Numeric scalar)
The model's temporal buffer capacity.}

\item{buffer_length_items}{(Integerish scalar)
The model's itemwise buffer capacity.}

\item{only_learn_from_buffer}{(Logical scalar)
If TRUE, then n-grams are only learned if they fit within
the memory buffer.}

\item{only_predict_from_buffer}{(Logical scalar)
If TRUE, then the context used for prediction is limited by the memory buffer.
Specifically, for a context to be used for prediction,
the first symbol within that context must still be within the buffer
at the point immediately before the predicted event occurs.}
}
\value{
A PPM-decay model object. 
These objects have reference semantics.
}
\description{
Creates a decay-based PPM model.
}
\details{
Decay-based PPM models generalise the PPM algorithm to incorporate
memory decay, where the effective counts of observed n-grams
decrease over time to reflect processes of auditory memory.

The model implements interpolated smoothing with escape method A, 
and explicitly disables exclusion and update exclusion.
}
\seealso{
\code{\link{new_ppm_simple}},
\code{\link{model_seq}}.
}
