% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm-package.R
\docType{package}
\name{ppm-package}
\alias{ppm}
\alias{ppm-package}
\title{ppm: What the Package Does in One 'Title Case' Line}
\description{
One paragraph description of what the package does as one or more full sentences.
}
\keyword{internal}
