% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{new_ppm_decay}
\alias{new_ppm_decay}
\title{Create decay-based PPM model}
\usage{
new_ppm_decay(alphabet_size, order_bound = 10L, buffer_length_time = 2,
  buffer_length_items = 13L, buffer_weight = 0.77,
  only_learn_from_buffer = TRUE, stm_half_life = 2,
  stm_weight = 0.53, ltm_weight = 0, noise = 0.5)
}
\arguments{
\item{alphabet_size}{(Integerish scalar)}

\item{order_bound}{(Integerish scalar)}

\item{buffer_length_time}{(Numeric scalar)}

\item{buffer_length_items}{(Integerish scalar)}

\item{buffer_weight}{(Numeric scalar)}

\item{only_learn_from_buffer}{(Logical scalar)
If TRUE, then n-grams are only learned if they fit within
the memory buffer.}

\item{stm_half_life}{(Numeric scalar)}

\item{stm_weight}{(Numeric scalar)}

\item{ltm_weight}{(Numeric scalar)}

\item{noise}{(Numeric scalar)}
}
\value{
A PPM-decay model object. 
These objects have reference semantics.
}
\description{
Creates a decay-based PPM model.
}
\details{
Decay-based PPM models generalise the PPM algorithm to incorporate
memory decay, where the effective counts of observed n-grams
decrease over time to reflect processes of auditory memory.

The model implements interpolated smoothing with escape method A, 
and explicitly disables exclusion and update exclusion.
}
\seealso{
\code{\link{new_ppm_simple}},
\code{\link{model_seq}}.
}
