% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{new_ppm_decay}
\alias{new_ppm_decay}
\title{Create decay-based PPM model}
\usage{
new_ppm_decay(alphabet_size, order_bound = 10L, buffer_length_time = 2,
  buffer_length_items = 13L, buffer_weight = 0.77,
  only_learn_from_buffer = TRUE, only_predict_from_buffer = TRUE,
  stm_half_life = 2, stm_weight = 0.53, ltm_weight = 0,
  noise = 0.5)
}
\arguments{
\item{alphabet_size}{(Integerish scalar)
The size of the alphabet from which sequences are drawn.}

\item{order_bound}{(Integerish scalar)
The model's Markov order bound.}

\item{buffer_length_time}{(Numeric scalar)
The model's temporal buffer capacity.}

\item{buffer_length_items}{(Integerish scalar)
The model's itemwise buffer capacity.}

\item{buffer_weight}{(Numeric scalar)
The weight of n-grams within the buffer.}

\item{only_learn_from_buffer}{(Logical scalar)
If TRUE, then n-grams are only learned if they fit within
the memory buffer.}

\item{only_predict_from_buffer}{(Logical scalar)
If TRUE, then the context used for prediction is limited by the memory buffer.
Specifically, for a context to be used for prediction,
the first symbol within that context must still be within the buffer
at the point immediately before the predicted event occurs.}

\item{stm_half_life}{(Numeric scalar)
Half life of the exponential-decay phase.}

\item{stm_weight}{(Numeric scalar)
The weight of n-grams immediately after leaving the buffer.}

\item{ltm_weight}{(Numeric scalar)
The weight of n-grams as time approaches infinity.}

\item{noise}{(Numeric scalar)
Scale parameter for the retrieval noise distribution.}
}
\value{
A PPM-decay model object. 
These objects have reference semantics.
}
\description{
Creates a decay-based PPM model.
}
\details{
Decay-based PPM models generalise the PPM algorithm to incorporate
memory decay, where the effective counts of observed n-grams
decrease over time to reflect processes of auditory memory.

The model implements interpolated smoothing with escape method A, 
and explicitly disables exclusion and update exclusion.
}
\seealso{
\code{\link{new_ppm_simple}},
\code{\link{model_seq}}.
}
