% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{new_ppm_simple}
\alias{new_ppm_simple}
\title{Create simple PPM model}
\usage{
new_ppm_simple(alphabet_size, order_bound = 10L,
  shortest_deterministic = TRUE, exclusion = TRUE,
  update_exclusion = TRUE, escape = "c", debug_smooth = FALSE)
}
\arguments{
\item{alphabet_size}{(Integerish scalar)
The size of the alphabet upon which the model will be trained and tested.}

\item{order_bound}{(Integerish scalar)
The model's Markov order bound. For example, an order bound of two means
that the model makes predictions based on the two preceding symbols.}

\item{shortest_deterministic}{(Logical scalar)
If TRUE, the model will 'select' the shortest available order
that provides a deterministic prediction, if such an order exists,
otherwise defaulting to the longest available order.
For a given prediction, if this rule results in a lower model order
than would have otherwise been selected, 
then full counts (not update-excluded counts) will be used for 
the highest model order (but not for lower model orders).
This behaviour matches the implementations of PPM* in 
\insertCite{Pearce2005;textual}{ppm} and
\insertCite{Bunton1996;textual}{ppm}.}

\item{exclusion}{(Logical scalar)
If TRUE, implements exclusion as defined in 
\insertCite{Pearce2005;textual}{ppm} and
\insertCite{Bunton1996;textual}{ppm}.}

\item{update_exclusion}{(Logical scalar)
If TRUE, implements update exclusion as defined in 
\insertCite{Pearce2005;textual}{ppm} and
\insertCite{Bunton1996;textual}{ppm}.}

\item{escape}{(Character scalar)
Takes values 'a', 'b', 'c', 'd', or 'ax',
corresponding to the eponymous escape methods 
in \insertCite{Pearce2005;textual}{ppm}.
Note that there is a mistake in the definition of escape method 
"AX" in \insertCite{Pearce2005;textual}{ppm}; 
the denominator of lambda needs to have 1 added.
This is what we implement here. Note that Pearce's
LISP implementation correctly adds 1 here, like us.}

\item{debug_smooth}{(Logical scalar)
Whether to print (currently rather messy and ad hoc) debug output
for smoothing.}
}
\value{
A PPM model object. 
These objects have reference semantics.
}
\description{
Creates a simple PPM model, that is, a PPM model
without any non-traditional features such as memory decay.
}
\note{
The implementation does not scale well to very large order bounds (> 50).
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{new_ppm_decay}},
\code{\link{model_seq}}.
}
