% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm-package.R
\docType{package}
\name{ppm-package}
\alias{ppm}
\alias{ppm-package}
\title{ppm: Prediction by Partial Matching}
\description{
Implements variants of the Prediction by Partial Matching algorithm.
}
\keyword{internal}
